pyproject.toml
[tool.poetry]
name = "moda-backend"
version = "0.1.0"
description = ""
authors = ["Sihyun Kim <k4sud0n@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.1.5"
django-ninja = "^1.3.0"
black = "^24.10.0"
requests = "^2.32.3"
django-ninja-jwt = "^5.3.5"
python-dotenv = "^1.0.1"
aiohttp = "^3.11.11"
gunicorn = "^23.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
